refactor: 패키지 재구조화
