+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,7 +4,9 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.controller.HandlerAdapter;
 import nextstep.mvc.controller.HandlerAdapterRegistry;
+import nextstep.mvc.controller.HandlerMapping;
 import nextstep.mvc.controller.HandlerMappingRegistry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
