+++ mvc/src/main/java/nextstep/mvc/controller/HandlerAdapter.java
@@ -1,9 +1,8 @@
-package nextstep.mvc;
-
-import nextstep.mvc.view.ModelAndView;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.view.ModelAndView;
 
 public interface HandlerAdapter {
     boolean supports(Object handler);
