+++ mvc/src/main/java/nextstep/mvc/controller/HandlerMapping.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
