+++ mvc/src/main/java/nextstep/mvc/controller/HandlerMappingRegistry.java
@@ -4,7 +4,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import nextstep.mvc.HandlerMapping;
 
 public class HandlerMappingRegistry {
