+++ mvc/src/main/java/nextstep/mvc/controller/RequestMappingHandlerMapping.java
@@ -3,7 +3,6 @@ package nextstep.mvc.controller;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
-import nextstep.mvc.HandlerMapping;
 import nextstep.web.support.RequestMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
