+++ mvc/src/main/java/nextstep/mvc/view/View.java
@@ -2,9 +2,9 @@ package nextstep.mvc.view;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-
 import java.util.Map;
 
 public interface View {
+
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
 }
