+++ mvc/src/test/java/nextstep/mvc/RequestMappingHandlerAdapterTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.stream.Stream;
 import nextstep.mvc.controller.HandlerExecution;
+import nextstep.mvc.controller.RequestMappingHandlerAdapter;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
